Create Julie's Hiring Store - Version 1.6
In version 6, as soon as an error is encountered, a warning is shown, and then the code continues to check for the next input. This results in multiple warning messages, which can be annoying to users. Then I started to think of I can created single warming message that shows all the errors encountered, and I created a list for all the error messages and uses "join" to are present all issues at once, making it easier for the user to correct their inputs.