Create Julie's Hired Store - Version 2.4
In version 2.4, I made some debugging adjustments to the delete function. Since I added an extra entry for the last name, the detail list for the delete function also needed to be changed. Additionally, I added an error message that appears when the inputted receipt number does not match any existing hire details.